+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  * `FileAppendTransaction` chunk size should default to 4096
 
+### Fixed
+
+ * `FileAppendTransaction.setTransactionId()` not locking the transaction ID list
+ * `TopicMessageSubmitTransaction.setTransactionId()` not locking the transaction ID list
+
 ### Deprecated
 
  * `AccountBalance.tokens` - Use the mirror node API https://docs.hedera.com/guides/docs/mirror-node-api/rest-api#api-v1-accounts instead